Split buyer seller protocol classes. Add delegate protocol for cancel trade
