+++ core/src/main/java/bisq/core/trade/protocol/BuyerAsTakerProtocol.java
@@ -97,6 +97,10 @@ public class BuyerAsTakerProtocol extends TradeProtocol implements BuyerProtocol
         }
     }
 
+    @Override
+    protected BuyersCancelTradeProtocol createCancelTradeProtocol() {
+        return new BuyersCancelTradeProtocol(trade);
+    }
 
     ///////////////////////////////////////////////////////////////////////////////////////////
     // Mailbox
