+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer/cancel/ProcessCancelTradeRequestRejectedMessage.java
@@ -15,11 +15,11 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.tasks.cancel;
+package bisq.core.trade.protocol.tasks.buyer.cancel;
 
-import bisq.core.trade.SellerTrade;
+import bisq.core.trade.BuyerTrade;
 import bisq.core.trade.Trade;
-import bisq.core.trade.messages.RequestCancelTradeMessage;
+import bisq.core.trade.messages.CancelTradeRequestRejectedMessage;
 import bisq.core.trade.protocol.tasks.TradeTask;
 import bisq.core.util.Validator;
 
@@ -30,9 +30,9 @@ import lombok.extern.slf4j.Slf4j;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
-public class ProcessRequestCancelTradeMessage extends TradeTask {
+public class ProcessCancelTradeRequestRejectedMessage extends TradeTask {
     @SuppressWarnings({"unused"})
-    public ProcessRequestCancelTradeMessage(TaskRunner taskHandler, Trade trade) {
+    public ProcessCancelTradeRequestRejectedMessage(TaskRunner taskHandler, Trade trade) {
         super(taskHandler, trade);
     }
 
@@ -40,18 +40,14 @@ public class ProcessRequestCancelTradeMessage extends TradeTask {
     protected void run() {
         try {
             runInterceptHook();
-            RequestCancelTradeMessage message = (RequestCancelTradeMessage) processModel.getTradeMessage();
+            CancelTradeRequestRejectedMessage message = (CancelTradeRequestRejectedMessage) processModel.getTradeMessage();
             Validator.checkTradeId(processModel.getOfferId(), message);
             checkNotNull(message);
 
-            processModel.getTradingPeer().setCanceledTradePayoutTxSignature(checkNotNull(message.getTxSignature()));
-
             // update to the latest peer address of our peer if the message is correct
             trade.setTradingPeerNodeAddress(processModel.getTempTradingPeerNodeAddress());
+            trade.setBuyersCancelTradeState(BuyerTrade.CancelTradeState.RECEIVED_REJECTED_MSG);
             processModel.removeMailboxMessageAfterProcessing(trade);
-
-            trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.RECEIVED_REQUEST);
-
             complete();
         } catch (Throwable t) {
             failed(t);
