+++ core/src/main/java/bisq/core/trade/protocol/tasks/seller/cancel/SendCancelTradeRequestRejectedMessage.java
@@ -15,38 +15,32 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.tasks.cancel;
+package bisq.core.trade.protocol.tasks.seller.cancel;
 
 import bisq.core.trade.SellerTrade;
 import bisq.core.trade.Trade;
-import bisq.core.trade.messages.CancelTradeRequestAcceptedMessage;
+import bisq.core.trade.messages.CancelTradeRequestRejectedMessage;
 import bisq.core.trade.messages.TradeMessage;
 import bisq.core.trade.protocol.tasks.SendMailboxMessageTask;
 
 import bisq.common.taskrunner.TaskRunner;
 
-import org.bitcoinj.core.Transaction;
-
 import java.util.UUID;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
 
 @Slf4j
-public class SendCancelTradeRequestAcceptedMessage extends SendMailboxMessageTask {
+public class SendCancelTradeRequestRejectedMessage extends SendMailboxMessageTask {
     @SuppressWarnings({"unused"})
-    public SendCancelTradeRequestAcceptedMessage(TaskRunner taskHandler, Trade trade) {
+    public SendCancelTradeRequestRejectedMessage(TaskRunner taskHandler, Trade trade) {
         super(taskHandler, trade);
     }
 
     @Override
     protected TradeMessage getMessage(String id) {
-        Transaction payoutTx = checkNotNull(trade.getPayoutTx(), "trade.getPayoutTx() must not be null");
-        return new CancelTradeRequestAcceptedMessage(
+        return new CancelTradeRequestRejectedMessage(
                 id,
-                payoutTx.bitcoinSerialize(),
                 processModel.getMyNodeAddress(),
                 UUID.randomUUID().toString()
         );
@@ -54,34 +48,28 @@ public class SendCancelTradeRequestAcceptedMessage extends SendMailboxMessageTas
 
     @Override
     protected void setStateSent() {
-        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_ACCEPTED_MSG_SENT);
+        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_REJECTED_MSG_SENT);
     }
 
     @Override
     protected void setStateArrived() {
-        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_ACCEPTED_MSG_ARRIVED);
+        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_REJECTED_MSG_ARRIVED);
     }
 
     @Override
     protected void setStateStoredInMailbox() {
-        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_ACCEPTED_MSG_IN_MAILBOX);
+        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_REJECTED_MSG_IN_MAILBOX);
     }
 
     @Override
     protected void setStateFault() {
-        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_ACCEPTED_MSG_SEND_FAILED);
+        trade.setSellersCancelTradeState(SellerTrade.CancelTradeState.REQUEST_REJECTED_MSG_SEND_FAILED);
     }
 
     @Override
     protected void run() {
         try {
             runInterceptHook();
-
-            if (trade.getPayoutTx() == null) {
-                log.error("trade.getPayoutTx() = " + trade.getPayoutTx());
-                failed("PayoutTx is null");
-                return;
-            }
 
             super.run();
         } catch (Throwable t) {
